requirements.txt
aiosqlite==0.19.0
alembic==1.11.1
anyio==3.7.0
apprise==1.4.0
asgi-lifespan==2.1.0
async-generator==1.10
asyncpg==0.27.0
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
coolname==2.2.0
croniter==1.3.15
cryptography==40.0.2
dateparser==1.1.8
docker==6.1.2
exceptiongroup==1.1.1
fastapi==0.95.2
fsspec==2023.5.0
google-api-core==2.11.0
google-api-python-client==2.87.0
google-auth==2.19.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
griffe==0.29.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.2
httplib2==0.22.0
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.17.3
kubernetes==26.1.0
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
oauthlib==3.2.2
orjson==3.8.14
outcome==1.2.0
packaging==23.1
pathspec==0.11.1
pendulum==2.1.2
prefect==2.10.11
prefect-gcp==0.4.2
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.8
Pygments==2.15.1
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
readchar==4.0.5
regex==2023.5.5
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.3.5
rsa==4.9
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
starlette==0.27.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typer==0.9.0
typing_extensions==4.6.2
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.22.0
webdriver-manager==3.8.6
websocket-client==1.5.2
websockets==11.0.3
wsproto==1.2.0
